+++ gpflow/__init__.py
@@ -17,6 +17,6 @@
 from __future__ import absolute_import
 from . import (likelihoods, kernels, ekernels, param,
                model, gpmc, sgpmc, priors, gpr, svgp,
-               vgp, sgpr, gplvm, tf_wraps, tf_hacks)
+               vgp, sgpr, gplvm, tf_wraps, tf_hacks, derivative_kernel)
 from ._version import __version__
 from ._settings import settings
